+++ config/run.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+gunicorn -w 3 -b 0.0.0.0:$PORT jsongs:app